Update the readme in the face of experience.
You might be wondering about the change to the new app link. It's a funny story. We used that link just now, and although it took us to the right place no problem, clicking "save" resulted in an infinite loop! It brought us rivght back to the same app registration page and there was no way to break out. But going to /settings/apps, and clicking the new app button avoided that situation. So. There.